+++ chat_be/src/main/java/com/hine/chat_be/repository/ConversationMemberRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConversationMemberRepository extends JpaRepository<ConversationMember, Integer> {
+public interface ConversationMemberRepository extends JpaRepository<ConversationMember, Long> {
 }
